requirements/base.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
alembic
pymysql
cryptography
python-jose[cryptography]
bcrypt
python-multipart
